feat(task): 滴答 CSV 导入、时段多提醒与 RPC 超时三档
支持栖息地数据维护导入滴答备份；任务 start_at/多提醒持久化；读/写/长任务超时 3/10/30s，大批量导入 10min；修复超大 sort_order 拖垮提醒调度。

+++ src/host/core/db/pg/entity/search/entity-search-repo.ts
@@ -41,7 +41,7 @@ function defaultOrderBy(primary_component?: string, opts?: { hasQuery?: boolean 
     (primary_component === TASK_ITEM_COMPONENT || primary_component === CONTENT_BLOCK_COMPONENT) &&
     !opts?.hasQuery
   ) {
-    return [sql`COALESCE((${entities.body}->>'sort_order')::int, 0)`, asc(entities.id)] as const;
+    return [sql`COALESCE((${entities.body}->>'sort_order')::bigint, 0)`, asc(entities.id)] as const;
   }
   return [desc(entities.updated_at), asc(entities.id)] as const;
 }
